+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * bugfix - add context for entryLogger
 * language clarity
 * Spanish translation
+* add README dialog to app for detailed more info and link to project page
 
 ### 4.3.0 (2020-6-13)
 * added headset receiver to passive service for autoswitch inputs